refactor: staged_ledger_diff.rs with signed_command.rs & user_command.rs
